requirements.txt
openai>=1.0.0
gradio
SpeechRecognition
gtts
python-dotenv
